Split GenericTypeTest out of ProcessorTest
