+++ src/test/java/org/inferred/freebuilder/processor/GenericTypeTest.java
@@ -1,0 +1,75 @@
+/*
+ * Copyright 2014 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.inferred.freebuilder.processor;
+
+import org.inferred.freebuilder.FreeBuilder;
+import org.inferred.freebuilder.processor.util.testing.BehaviorTester;
+import org.inferred.freebuilder.processor.util.testing.SourceBuilder;
+import org.inferred.freebuilder.processor.util.testing.TestBuilder;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
+
+@RunWith(JUnit4.class)
+public class GenericTypeTest {
+
+  private final BehaviorTester behaviorTester = new BehaviorTester();
+
+  @Test
+  public void testGenericInterface() {
+    behaviorTester
+        .with(new Processor())
+        .with(new SourceBuilder()
+            .addLine("package com.example;")
+            .addLine("@%s", FreeBuilder.class)
+            .addLine("public interface DataType<A, B> {")
+            .addLine("  A getPropertyA();")
+            .addLine("  B getPropertyB();")
+            .addLine("")
+            .addLine("  public static class Builder<A, B> extends DataType_Builder<A, B> {}")
+            .addLine("}")
+            .build())
+        .with(new TestBuilder()
+            .addLine("com.example.DataType<Integer, Boolean> value =")
+            .addLine("    new com.example.DataType.Builder<Integer, Boolean>()")
+            .addLine("        .setPropertyA(11)")
+            .addLine("        .setPropertyB(true)")
+            .addLine("        .build();")
+            .addLine("assertEquals(11, (int) value.getPropertyA());")
+            .addLine("assertTrue(value.getPropertyB());")
+            .build())
+        .runTest();
+  }
+
+  @Test
+  public void testGenericInterface_compilesWithoutWarnings() {
+    behaviorTester
+        .with(new Processor())
+        .with(new SourceBuilder()
+            .addLine("package com.example;")
+            .addLine("@%s", FreeBuilder.class)
+            .addLine("public interface DataType<A, B> {")
+            .addLine("  A getPropertyA();")
+            .addLine("  B getPropertyB();")
+            .addLine("")
+            .addLine("  public static class Builder<A, B> extends DataType_Builder<A, B> {}")
+            .addLine("}")
+            .build())
+        .compiles()
+        .withNoWarnings();
+  }
+
+}
